Sprint 7b (#591 D2): split recruitment notice/adjutancy/reason/call repositories
Static read/write split (CustomFieldRepository pattern) for the four smaller
recruitment repositories — each gets a <Name>Reader + <Name>Writer behind the
existing class kept as a thin delegating façade (constants + @phpstan-type
retained; reader/writer share the repo's cache group; table-name accessor
duplicated; tech-debt note #563 B3). Method bodies moved verbatim.

Each repo test gains @covers for its Reader/Writer + a setUp preload so pcov
attributes the delegated coverage. Behavior-preserving — Recruitment suite green
(566 tests); PHPStan L8 + WPCS clean.
