+++ tests/Unit/RecruitmentAdjutancyRepositoryTest.php
@@ -15,6 +15,8 @@ use FreeFormCertificate\Recruitment\RecruitmentAdjutancyRepository;
  * (matérias) reused across notices.
  *
  * @covers \FreeFormCertificate\Recruitment\RecruitmentAdjutancyRepository
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentAdjutancyReader
+ * @covers \FreeFormCertificate\Recruitment\RecruitmentAdjutancyWriter
  */
 class RecruitmentRecruitmentAdjutancyRepositoryTest extends TestCase {
 
@@ -38,6 +40,12 @@ class RecruitmentRecruitmentAdjutancyRepositoryTest extends TestCase {
 		Functions\when( 'wp_cache_set' )->justReturn( true );
 		Functions\when( 'wp_cache_delete' )->justReturn( true );
 		Functions\when( 'current_time' )->justReturn( '2026-05-01 10:00:00' );
+
+		// pcov does not record lines for files first autoloaded mid-test-method,
+		// so the extracted reader/writer's coverage would attribute to nothing.
+		// Preload the classes here so pcov attributes their lines to this test.
+		class_exists( '\\FreeFormCertificate\\Recruitment\\RecruitmentAdjutancyReader' );
+		class_exists( '\\FreeFormCertificate\\Recruitment\\RecruitmentAdjutancyWriter' );
 
 		$this->wpdb->shouldReceive( 'prepare' )
 			->andReturnUsing(
